kg-intel/experiments/requirements.txt
neo4j>=5.0
pymongo>=4.0
python-dotenv
langchain-neo4j
langchain-openai
openai
